backend/requirements.txt
# CampusIQ Backend — Python Dependencies
# Compatible with Python 3.12+ | Fixes pydantic-core meta generation error
# Install: pip install -r requirements.txt

# ── Web Framework ──────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ── Database ───────────────────────────────────────
motor==3.6.0
pymongo==4.9.2

# ── Authentication ─────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# ── AI / RAG Pipeline ──────────────────────────────
langchain==0.3.14
langchain-community==0.3.14
langchain-huggingface==0.1.2
langchain-groq==0.2.3
chromadb==0.6.3
sentence-transformers==3.3.1
groq==0.13.0

# ── Document Parsing ───────────────────────────────
pdfplumber==0.11.4
python-docx==1.1.2
docx2txt==0.8

# ── Translation ────────────────────────────────────
deep-translator==1.11.4

# ── Notifications (Optional) ───────────────────────
firebase-admin==6.6.0

# ── Security ───────────────────────────────────────
cryptography==43.0.3
python-dotenv==1.0.1

# ── Validation ─────────────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0

# ── Rate Limiting ──────────────────────────────────
slowapi==0.1.9

# ── Async ──────────────────────────────────────────
httpx==0.28.1
aiofiles==24.1.0
